Sensor/requirements.txt
# Python packages required for Sensor Raspberry
numpy
paho-mqtt
RPi.GPIO
